Swap float64 for float32 throughout, fixes mlp layer by having two, one for edge, one for nodes. Fixes feature length

+++ kglib/kgcn_experimental/encode.py
@@ -110,4 +110,4 @@ class TypeEncoder(snt.AbstractModule):
 
 
 def pass_input_through_op(op):
-    return lambda features: tf.concat([tf.expand_dims(features[:, 0], axis=1), op(features[:, 1:])], axis=1)
+    return lambda features: tf.concat([tf.expand_dims(tf.cast(features[:, 0], dtype=tf.float32), axis=1), op(features[:, 1:])], axis=1)
